List live added fields on hot-reload --status
Agents could not see which added fields were still live because
--status only listed methods. AddedField rows and AddedFieldTotal
come from the existing ledger and stay out of ActivePatchTotal.
