+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadStatusExecutor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f1bd0311e4c304aad9a47afdde54e9d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
